// Develop/src/interfaces/Candidate.interface.tsx
// TODO: Create an interface for the Candidate objects returned by the API
// Expected stuff: name, username, location, avatar, email, html_url, and company
export interface Candidate {
    id: number;
    name: string;
    username: string;
    location?: string;
    avatar_url: string;
    email?: string;
    html_url: string;
    company?: string;
}